refactor(testutils): host SkipIfGoTip in the shared package
The helper first landed in its own leaf package to avoid pulling the
heavier instrumentation/testutils package (and its go-libddwaf, appsec,
and tracer imports) into the gotesting reflection test binary. That
premise was wrong: go list -test -deps shows the gotesting test binary
already links go-libddwaf through the tracer and civisibility chains, so
hosting the helper in instrumentation/testutils adds no new dependency
and reuses the established cross-module test-helper package instead of
adding a new one.
