+++ instrumentation/testutils/goversion_test.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2026 Datadog, Inc.
 
-package goversion
+package testutils
 
 import (
 	"runtime"
